another commit

+++ struct.h
@@ -3,5 +3,5 @@ struct Record
 {
     bool type;// 1 - запись о продаже; 0 - запись о закупке
     unsigned amount; // количество товара
-    unsigned cost; // цена за штуку
+    double cost; // цена за штуку
 };
